Set up logging config and added logs

+++ chat/consumers.py
@@ -1,11 +1,15 @@
+import logging
 import json
 from asgiref.sync import async_to_sync
 from channels.generic.websocket import WebsocketConsumer
 from channels.exceptions import DenyConnection
 from django.db.utils import IntegrityError
+from django.conf import settings
 import chat.models.channel as Channel
 import chat.models.message as Message
 from chat.constants import MESSAGE, PREFIX
+
+logger = logging.getLogger(__name__)
 
 class ChatConsumer(WebsocketConsumer):
     """Custom WebsocketConsumer for handling chat web socket requests
@@ -20,11 +24,15 @@ class ChatConsumer(WebsocketConsumer):
     def connect(self):
         session = self.scope['session']
         if session.session_key is None:
-            # for local development
-            session.create()
+            if settings.DEBUG is True:    # for local development
+                session.create()
+            else:
+                logger.error('SuspiciousOperation : session not created in production')
+                raise DenyConnection
 
         # room_id in URL comes only in group chat
         if 'room_id' in self.scope['url_route']['kwargs']:
+            self.is_group_consumer = True
             self.room_id = self.scope['url_route']['kwargs']['room_id']
             try:
                 new_channel = Channel.GroupChannel.objects.create(
@@ -33,6 +41,7 @@ class ChatConsumer(WebsocketConsumer):
                     group_room_id = self.room_id
                 )
             except IntegrityError as excp:
+                logger.error('Non-existing room id %d', self.room_id)
                 raise DenyConnection from excp
             async_to_sync(self.channel_layer.group_add)(
                 PREFIX.GROUP_ROOM + str(self.room_id),
@@ -49,7 +58,7 @@ class ChatConsumer(WebsocketConsumer):
                     }
                 }
             )
-            self.is_group_consumer = True
+            logger.info('New group channel created with room_id %d', self.room_id)
         else:
             new_channel = Channel.IndividualChannel.objects.create(
                 name=self.channel_name,
@@ -59,18 +68,24 @@ class ChatConsumer(WebsocketConsumer):
                 PREFIX.INDIVIDUAL_CHANNEL + str(new_channel.id),
                 self.channel_name
             )
+            logger.info('New individual channel created')
         self.channel_id = new_channel.id
+        logger.info('Channel id: %d', self.channel_id)
         self.accept()
 
     def disconnect(self, code):
+        if self.channel_id is None:
+            return
         if self.is_group_consumer:
             Channel.GroupChannel.objects.filter(pk=self.channel_id).delete()
+            logger.info('Group channel deleted')
         else:
             Channel.IndividualChannel.objects.filter(pk=self.channel_id).delete()
             async_to_sync(self.channel_layer.group_discard)(
                 PREFIX.INDIVIDUAL_CHANNEL + str(self.channel_id),
                 self.channel_name
             )
+            logger.info('Individual channel deleted')
         if self.room_id is not None:
             group_prefix = PREFIX.GROUP_ROOM if self.is_group_consumer else PREFIX.INDIVIDUAL_ROOM
             async_to_sync(self.channel_layer.group_discard)(
@@ -88,6 +103,8 @@ class ChatConsumer(WebsocketConsumer):
                     #TODO: send user info who has left
                 }
             )
+            logger.info("Room id: %d, Channel id: %d", self.room_id, self.channel_id)
+
 
     def receive(self, text_data=None, bytes_data=None):
         payload_json = json.loads(text_data)
@@ -95,8 +112,13 @@ class ChatConsumer(WebsocketConsumer):
         message_data = payload_json['data']
         if message_type == MESSAGE.TEXT:
             if self.room_id is None:
+                logger.error('SuspiciousOperation : message received outside of room')
                 self.close()
                 return
+            logger.info('Text message received in room id %d', self.room_id)
+            # TODO: remove this log as messages will be encrypted
+            logger.info('%s', message_data['text'])
+            # TODO: log sender info
             group_prefix = PREFIX.INDIVIDUAL_ROOM
             if self.is_group_consumer:
                 Message.TextMessage.objects.create(
